Statement, Batch: Add per statement exec profile
Statements and Batches are the same in terms of execution profiles, so the
below covers both of them.

Due to the following limitations of CPP driver API:
- `cass_statement_set_execution_profile()` does not have access to exec
profiles map stored in CassSession,
- `cass_session_execute()` can't mutate the shared statement, but only
its own copy of the statement,

exec profiles in statements must be split into names and handles.
To this end, PerStatementExecProfile enum is used. It covers both
already resolved profile handles and yet-to-be-resolved profile names.
Names are set eagerly on `cass_statement_set_execution_profile()` call.
Note: CPP driver resolves the exec profile's name in the global hashmap
on each query. As we want to avoid it, we resolve it only once, lazily,
on the first query after the name it set. To this end, RwLock is used.
The reason why the handle can't be set eagerly is, again, that
the mentioned function does not provide any access to the CassSession.